panel/backend/pyproject.toml
[project]
name = "lean-panel-backend"
version = "0.1.0"
description = "Lean control panel backend (FastAPI)"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "httpx>=0.26",
  "pyjwt[crypto]>=2.9",
  "cryptography>=42",
  "passlib[argon2]>=1.7.4",
  "python-multipart>=0.0.9",
  "jinja2>=3.1",
]

[tool.uvicorn]
factory = false

[tool.pytest.ini_options]
testpaths = ["tests"]
